Major overhaul
Moved project to Eclipse, makes it easier to do stuff. Also moved from
using JFrame etc to using LWJGL which is going to help in the long run
for doing things, however, generating large grids takes a little too
long and will have to be sorted soon. (maybe in the next update)